renamed services to be more consistent
