+++ santa/AngularClient/santa-pone-central/src/app/admin-control/survey-control/survey-control.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { Survey, Question } from 'src/classes/survey';
 import { Validators, FormControl, FormBuilder, FormGroup } from '@angular/forms';
-import { SantaApiGetService, SantaApiPutService, SantaApiPostService, SantaApiDeleteService } from 'src/app/services/santaApiService.service';
-import { MapResponse, MapService } from 'src/app/services/mapService.service';
+import { SantaApiGetService, SantaApiPutService, SantaApiPostService, SantaApiDeleteService } from 'src/app/services/santa-api.service';
+import { MapResponse, MapService } from 'src/app/services/mapper.service';
 import { GathererService } from 'src/app/services/gatherer.service';
 import { Client } from 'src/classes/client';
 import { ThrowStmt } from '@angular/compiler';
