+++ santa/AngularClient/santa-pone-central/src/app/headquarters/manual-add/manual-add.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, ViewChildren, QueryList, Output, EventEmitter } from '@angular/core';
-import { SantaApiGetService, SantaApiPostService } from 'src/app/services/santaApiService.service';
+import { SantaApiGetService, SantaApiPostService } from 'src/app/services/santa-api.service';
 import { GathererService } from 'src/app/services/gatherer.service';
-import { MapService, MapResponse } from 'src/app/services/mapService.service';
+import { MapService, MapResponse } from 'src/app/services/mapper.service';
 import { CountriesService } from 'src/app/services/countries.service';
 import { FormBuilder, FormGroup, Validators, FormControl } from '@angular/forms';
 import { SurveyFormComponent } from 'src/app/signup/survey-form/survey-form.component';
