+++ santa/AngularClient/santa-pone-central/src/app/profile/information/information.component.ts
@@ -1,9 +1,9 @@
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { Profile } from 'src/classes/profile';
-import { ProfileService } from 'src/app/services/Profile.service';
+import { ProfileService } from 'src/app/services/profile.service';
 import { EventType } from 'src/classes/eventType';
 import { ChangeSurveyResponseModel, ClientAddressResponse } from 'src/classes/responseTypes';
-import { SantaApiPutService } from 'src/app/services/santaApiService.service';
+import { SantaApiPutService } from 'src/app/services/santa-api.service';
 import { SurveyResponse, Survey } from 'src/classes/survey';
 import { FormGroup, FormBuilder, FormControl, Validators } from '@angular/forms';
 import { CountriesService } from 'src/app/services/countries.service';
